Add new output format: "signify"
This is BSD tool used to sign release, with simpler interface and
without support of old key types. It does not do encryption.

This is how signify can be used:

  $ signify -G -n -p new.pub -s new.key
  $ signify -S -s new.key -m message.txt # creates message.txt.sig
  $ signify -V -p new.pub -m message.txt # checks signature

Now first step can be replaced with:

  $ passphrase2pgp -f signify >new.combined

First two lines are the public key, and the next two are the secret key.
Note: signify(1) insists each begins with "untrusted comment: ".

Co-Authored: Christopher Wellons <wellons@nullprogram.com>